test: complete feature suite migration (Group C)
Adds 10 new feature_*.rs files covering the remaining test surface,
then deletes all 15 e2e_*.rs files and removes every mention of the
e2e suite from the codebase.

New files (76 tests total):
  feature_doc_lifecycle       — didClose, didSave, willSaveWaitUntil, didChange
  feature_formatting          — document/range/on-type formatting
  feature_document_link       — require paths, @link docblocks
  feature_file_ops            — willCreate/Rename/DeleteFiles, PSR-4 stubs
  feature_semantic_tokens     — full, range, delta, delta-fallback
  feature_incremental         — cache invalidation, cross-file republish, burst
  feature_configuration       — didChangeConfiguration, PHP version, refresh
  feature_server              — initialize, shutdown, protocol stubs, concurrency
  feature_workspace_scan      — CREATED/CHANGED/DELETED events, excludePaths
  feature_workspace_folders   — add/remove folders, didCreate/Rename/DeleteFiles

Cleanup:
  - Rename test fn references_fast_path_final_class_cross_file_e2e (drop suffix)
  - Update doc comments in common/ from "E2E" to neutral language
  - Update fixtures/README.md, benches/README.md, docs/salsa-migration.md
  - Delete TEST_SUITE_PLAN.md (migration complete)
